Refactor event stock event listener into queue processor
